+++ apollo-router/src/plugins/telemetry/metrics/otlp.rs
@@ -6,9 +6,9 @@ use tower::BoxError;
 use crate::metrics::aggregation::MeterProviderType;
 use crate::plugins::telemetry::config::Conf;
 use crate::plugins::telemetry::metrics::CustomAggregationSelector;
-use crate::plugins::telemetry::metrics::MetricsConfigurator;
 use crate::plugins::telemetry::otlp::TelemetryDataKind;
-use crate::plugins::telemetry::reload::builder::MetricsBuilder;
+use crate::plugins::telemetry::reload::metrics::MetricsBuilder;
+use crate::plugins::telemetry::reload::metrics::MetricsConfigurator;
 
 impl MetricsConfigurator for super::super::otlp::Config {
     fn config(conf: &Conf) -> &Self {
