+++ apollo-router/src/plugins/telemetry/metrics/prometheus.rs
@@ -16,8 +16,8 @@ use crate::ListenAddr;
 use crate::metrics::aggregation::MeterProviderType;
 use crate::plugins::telemetry::config::Conf;
 use crate::plugins::telemetry::metrics::CustomAggregationSelector;
-use crate::plugins::telemetry::metrics::MetricsConfigurator;
-use crate::plugins::telemetry::reload::builder::MetricsBuilder;
+use crate::plugins::telemetry::reload::metrics::MetricsBuilder;
+use crate::plugins::telemetry::reload::metrics::MetricsConfigurator;
 use crate::services::router;
 
 /// Prometheus configuration
