+++ apollo-router/src/plugins/telemetry/tracing/otlp.rs
@@ -8,7 +8,7 @@ use tower::BoxError;
 use crate::plugins::telemetry::config::Conf;
 use crate::plugins::telemetry::otel::named_runtime_channel::NamedTokioRuntime;
 use crate::plugins::telemetry::otlp::TelemetryDataKind;
-use crate::plugins::telemetry::reload::builder::TracingBuilder;
+use crate::plugins::telemetry::reload::tracing::TracingBuilder;
 use crate::plugins::telemetry::tracing::SpanProcessorExt;
 use crate::plugins::telemetry::tracing::TracingConfigurator;
