+++ apollo-router/src/plugins/telemetry/tracing/zipkin.rs
@@ -12,7 +12,7 @@ use crate::plugins::telemetry::config::Conf;
 use crate::plugins::telemetry::config::GenericWith;
 use crate::plugins::telemetry::endpoint::UriEndpoint;
 use crate::plugins::telemetry::otel::named_runtime_channel::NamedTokioRuntime;
-use crate::plugins::telemetry::reload::builder::TracingBuilder;
+use crate::plugins::telemetry::reload::tracing::TracingBuilder;
 use crate::plugins::telemetry::tracing::BatchProcessorConfig;
 use crate::plugins::telemetry::tracing::SpanProcessorExt;
 use crate::plugins::telemetry::tracing::TracingConfigurator;
